PDF export seems to expose possibly restricted content to an external service #30 * The data sent by the browser is compressed so we need to pre-process it before parsing. * Refactor the code a bit in order to be able to test it more easily.

+++ application-diagram-api/src/main/java/com/xwiki/diagram/internal/DiagramExportRequestFactory.java
@@ -148,7 +148,7 @@ public class DiagramExportRequestFactory
     {
         String value = request.getParameter("scale");
         if (value != null) {
-            diagramExportRequest.scale = Float.parseFloat(value);
+            diagramExportRequest.scale = Double.parseDouble(value);
             if (diagramExportRequest.scale <= 0) {
                 throw new IllegalArgumentException("Invalid export scale.");
             }
